third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-display-contents-crash.html

<div style="display: contents"><a href="#"></a></div>
Shouldn't crash.
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
    test(() => {
        assert_true(!!window.testRunner);
        testRunner.overridePreference("WebKitTabToLinksPreferenceKey", 1);
        testRunner.overridePreference("WebKitSpatialNavigationEnabled", 1);
        eventSender.keyDown("ArrowLeft");
    }, "No crash or assertion failure");
</script>
